@@ -1,33 +1,50 @@
 {
   "name": "@bjerk/eslint-config",
-  "version": "2.1.0",
+  "version": "3.0.1",
   "repository": "github.com:bjerkio/eslint-config",
-  "author": "Bjerk AS",
   "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public"
+  "author": "Bjerk AS",
+  "main": "index.js",
+  "prettier": "@cobraz/prettier",
+  "devDependencies": {
+    "@cobraz/prettier": "^2.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^8.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.0.4",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.0.1",
+    "eslint-plugin-unicorn": "^43.0.2",
+    "prettier": "^2.7.1"
+  },
+  "peerDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^7.32.0 || ^8.2.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-eslint-comments": "^3.2.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jest": "^27.0.4",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-promise": "^6.0.1",
+    "eslint-plugin-unicorn": "^43.0.2",
+    "prettier": "^2.7.1"
+  },
+  "peerDependenciesMeta": {
+    "prettier": {
+      "optional": true
+    }
   },
   "engines": {
     "node": ">=16"
   },
-  "main": "index.js",
-  "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.30.6",
-    "@typescript-eslint/parser": "^5.30.6",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.5.3",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-unicorn": "^43.0.1"
-  },
-  "devDependencies": {
-    "prettier": "^2.7.1"
-  },
   "volta": {
-    "node": "16.14.0",
-    "yarn": "1.22.17"
+    "node": "16.14.0"
+  },
+  "publishConfig": {
+    "access": "public"
   }
 }
